refactor(structure): directory구조를 변경한다
