+++ src/card-register/components/card-form/card-expired-period/ExpiredPeriodFieldset.tsx
@@ -1,0 +1,18 @@
+import React from 'react';
+import { FieldsetContent, Fieldset, FieldsetHead } from '../Fieldset';
+import ExpiredPeriodInputContainer from './ExpiredPeriodInputContainer';
+
+function ExpiredPeriodFieldset() {
+  return (
+    <Fieldset>
+      <FieldsetHead>
+        <label>만료일</label>
+      </FieldsetHead>
+      <FieldsetContent>
+        <ExpiredPeriodInputContainer />
+      </FieldsetContent>
+    </Fieldset>
+  );
+}
+
+export default ExpiredPeriodFieldset;
