+++ src/card-register/components/card/Card.tsx
