+++ src/card-register/constants.ts
