--- src/components/card-form/card-password/CardPasswordFieldset.tsx
@@ -1,18 +1,0 @@
-import React from 'react';
-import CardPasswordInputContainerList from '../card-password/CardPasswordInputContainerList';
-import { FieldsetContent, Fieldset, FieldsetHead } from '../Fieldset';
-
-function CardPasswordFieldset() {
-  return (
-    <Fieldset>
-      <FieldsetHead>
-        <label>카드 비밀번호</label>
-      </FieldsetHead>
-      <FieldsetContent>
-        <CardPasswordInputContainerList />
-      </FieldsetContent>
-    </Fieldset>
-  );
-}
-
-export default CardPasswordFieldset;
